+++ src/models.rs
@@ -1,5 +1,6 @@
 use serde::{Deserialize, Serialize};
 
+
 /// Filters for VisualMediaPicker.
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, Deserialize, Serialize)]
 pub enum VisualMediaTarget {
@@ -15,7 +16,6 @@ pub enum VisualMediaTarget {
 }
 
 /// The application specific directory.  
-// TODO: Add #[non_exhaustive] from version 2.0.0
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, Deserialize, Serialize)]
 pub enum PrivateDir {
 
@@ -41,7 +41,6 @@ pub enum PrivateDir {
 
 /// Directory in which to place images that are available to other applications and users.  
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, Deserialize, Serialize)]
-// TODO: Add #[non_exhaustive] from version 2.0.0
 pub enum PublicImageDir {
 
 	/// Standard directory in which to place pictures that are available to the user.  
@@ -56,7 +55,6 @@ pub enum PublicImageDir {
 }
 
 /// Directory in which to place videos that are available to other applications and users.  
-// TODO: Add #[non_exhaustive] from version 2.0.0
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, Deserialize, Serialize)]
 pub enum PublicVideoDir {
 
@@ -72,7 +70,6 @@ pub enum PublicVideoDir {
 }
 
 /// Directory in which to place audios that are available to other applications and users.  
-// TODO: Add #[non_exhaustive] from version 2.0.0
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, Deserialize, Serialize)]
 pub enum PublicAudioDir {
 
@@ -119,7 +116,6 @@ pub enum PublicAudioDir {
 }
 
 /// Directory in which to place files that are available to other applications and users.  
-// TODO: Add #[non_exhaustive] from version 2.0.0
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, Deserialize, Serialize)]
 pub enum PublicGeneralPurposeDir {