+++ CHANGES.md
@@ -1,3 +1,10 @@
+# Verstion 1.4.0
+- Add `AndroidFs::take_persistable_read_permission`.
+- Add `AndroidFs::take_persistable_write_permission`.
+- Add `convert_string_to_file_path`.
+- Add `convert_file_path_to_string`.
+- Update documents.
+
 # Version 1.3.2
 - Update documents.
 
@@ -6,7 +13,7 @@
 
 # Version 1.3.0
 - Deprecate `AndroidFs::open_file_writable`.
-- Add [`AndroidFs::create_file`].
+- Add `AndroidFs::create_file`.
 - Add `PrivateStorage::create_file`.
 - Update documents.